CM-9826 Update source code for Kentico 11 (#26)
* Copy files from Mvc-dev and merged changelog.md

* CM-9828 Update links in Readme.md

* Remove -unstable suffix in AssemblyInfo.cs files

* Change NuGet dependencies to correct version of Kentico.Libraries

* Change Kentico MVC NuGet dependencies to release version

* Change copyright year to 2017 in files

* Fix corrupted SLN file

* Update requirements in Readme.md

* Downgrade nuget.exe to version 3.3

* Add DancingGoatMvc 11 site export package

* Update DancingGoatMvc site export package

+++ src/Kentico.Ecommerce/Models/OrderItem.cs
@@ -30,13 +30,13 @@ namespace Kentico.Ecommerce
         /// <summary>
         /// Gets the total price of the order item in the main currency.
         /// </summary>
-        public decimal Subtotal => (decimal)OriginalOrderItem.OrderItemTotalPriceInMainCurrency;
+        public decimal Subtotal => OriginalOrderItem.OrderItemTotalPriceInMainCurrency;
 
 
         /// <summary>
         /// Gets the unit price of the order item.
         /// </summary>
-        public decimal UnitPrice => (decimal)OriginalOrderItem.OrderItemUnitPrice;
+        public decimal UnitPrice => OriginalOrderItem.OrderItemUnitPrice;
 
 
         /// <summary>
